Fixed Navbar urls to use local host rather than live url. Also added initial trial of testing page lang property.
