+++ csunplugged/tests/topics/in-browser/test_in_browser.py
@@ -64,20 +64,6 @@ class BrowserTest(unittest.TestCase):
         if "Printables" not in element:
             raise Exception("Unable to load local page!")
         driver.quit()
-    #
-    # def test_local_3(self):
-    #     driver = webdriver.Remote(
-    #         command_executor=COMMAND_EXECUTOR,
-    #         desired_capabilities=local_server)
-    #
-    #     driver.get("http://localhost/")
-    #     driver.find_element_by_xpath(
-    #         "(.//*[normalize-space(text()) and normalize-space(.)="
-    #         "'How do I teach CS Unplugged?'])[1]/following::h2[1]").click()
-    #     element = driver.title
-    #     if "Topics - CS Unplugged" not in element:
-    #         raise Exception("Unable to load local page!")
-    #     driver.quit()
 
     def test_local(self):
         driver = webdriver.Remote(
