+++ csunplugged/tests/topics/in-browser/test_navbar.py
@@ -1,41 +1,14 @@
-import os
 import unittest
 
 from django.test import tag
 from selenium import webdriver
 
-ACCESS_KEY = os.environ['KEY']
-COMMAND_EXECUTOR = 'http://cseducationresea1:' + ACCESS_KEY + '@hub.browserstack.com:80/wd/hub'
-
-local_server = {
-    'os': 'Windows',
-    'os_version': '10',
-    'browser': 'Chrome',
-    'browser_version': '63.0',
-    'resolution': '1920x1080',
-    'browserstack.local': 'true'
-}
-
-TITLE_ERROR_TEXT = "Failed to load page\nExpected title = {}, got {}"
-URL_ERROR_TEXT = "Failed to load page\nExpected url = {}, not {}"
-
-
-def check_title_and_url(driver, expected_title, expected_url):
-    if driver.title != expected_title:
-        raise Exception(TITLE_ERROR_TEXT.format(expected_title, driver.title))
-    if driver.current_url != expected_url:
-        raise Exception(URL_ERROR_TEXT.format(expected_url, driver.current_url))
-
-
-def load_home_page(driver):
-    driver.get("http://localhost/en")
-    if driver.title != "CS Unplugged":
-        raise Exception(TITLE_ERROR_TEXT.format(driver.title, "CS Unplugged"))
+from .helpers import *
 
 
 @tag("browser")
 class NavbarTest(unittest.TestCase):
-    """ Test cases for the in-browser test suite"""
+    """ Test cases for using the navigation bar"""
 
     def test_navbar_loads_pages(self):
         driver = webdriver.Remote(
@@ -51,11 +24,11 @@ class NavbarTest(unittest.TestCase):
 
         # Printables page
         driver.find_element_by_link_text("Printables").click()
-        check_title_and_url(driver, "Printables - CS Unplugged", "https://csunplugged.org/en/resources/")
+        check_title_and_url(driver, "Printables - CS Unplugged", "http://localhost/en/resources/")
 
         # About page
         driver.find_element_by_link_text("About").click()
-        check_title_and_url(driver, "About - CS Unplugged", "https://csunplugged.org/en/about/")
+        check_title_and_url(driver, "About - CS Unplugged", "http://localhost/en/about/")
 
         # Return to home page
         driver.find_element_by_id("navbar-brand-logo")
